errors repaired
